Handle case where no header is found
The header value is only assigned within a conditional, so when parsing sketchy ontologies that don't have properly written headers, it ends up giving an exception like `cannot access local variable 'header' where it is not associated with a value`. This PR changes that to send a warning instead and let the program move on. Alternatively, this could be a place to issue a proper exception